refactor: rename infrastructure/wdl to wdlindexer and fix bundle test artifact
- infrastructure/wdl collided with pkg/wdl and agents/tools/wdl, forcing
  an import alias at every use; the package is the WDL indexer, so name
  it that.
- The bundle not-found test wrote its OutputPath relative to the package
  directory, leaving a stray empty out/ dir in the source tree; use
  t.TempDir().
